Java 8 updates

+++ src/main/java/org/swrlapi/builtins/temporal/DatetimeStringProcessor.java
@@ -269,7 +269,7 @@ abstract class DatetimeStringProcessor
       int i = 1;
       while (i++ < this.gTokenIndex[granularity])
         tokenizer.nextToken();
-      return new Integer(tokenizer.nextToken()).intValue();
+      return new Integer(tokenizer.nextToken());
     } catch (Exception e) {
       Temporal.throwInvalidDatetimeStringException(datetimeString);
     }
